Adds basic app shell for generic feature pages
