+++ src/Apps/getAppRoutes.tsx
@@ -4,7 +4,8 @@ import { routes as artistRoutes } from "Apps/Artist/routes"
 import { routes as artworkRoutes } from "Apps/Artwork/routes"
 import { collectRoutes } from "Apps/Collect/collectRoutes"
 import { conversationRoutes } from "Apps/Conversation/routes"
-import { routes as featureRoutes } from "Apps/FeatureAKG/routes"
+import { routes as featureAKGRoutes } from "Apps/FeatureAKG/routes"
+import { routes as featureRoutes } from "Apps/Feature/routes"
 import { routes as identityVerificationRoutes } from "Apps/IdentityVerification/routes"
 import { routes as orderRoutes } from "Apps/Order/routes"
 import { routes as searchRoutes } from "Apps/Search/routes"
@@ -23,6 +24,9 @@ export function getAppRoutes(): RouteConfig[] {
     },
     {
       routes: conversationRoutes,
+    },
+    {
+      routes: featureAKGRoutes,
     },
     {
       routes: featureRoutes,
